% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_make.R
\name{bin_make}
\alias{bin_make}
\title{Generate the bin of observation time}
\usage{
bin_make(tp, bin_len)
}
\arguments{
\item{tp}{A set of observation time}

\item{bin_len}{The width of bin}
}
\value{
A list of different bins.
}
\description{
Break the observation time into different bins, with the width of the bin as bin_len
}
\examples{
tp <- c(1:5, 51:55, 101:105, 501:505, 601:605)
bin_len <- 50
bin_make(tp, bin_len)

}
