% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{lar_bootstrap}
\alias{lar_bootstrap}
\title{Nonparametric block bootstrapping for LAR models}
\usage{
lar_bootstrap(X, block_list, group_id = NULL, seed = NULL)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals within
study area, and the states or status of individuals
during each sampling period.}

\item{block_list}{A block list for a series of observation time. For example,
block_list = list(c(1:5), c(51:55), c(101:105), c(501:505), c(601:605)).}

\item{group_id}{Groups of individuals. If X is a list, please input group_id. If X is a matrix,
this parameter can be skipped and takes the default \code{NULL} value.}

\item{seed}{Random seed}
}
\value{
The block bootstrap samples.
}
\description{
Nonparametric block bootstrapping for LAR models
}
\details{
See block bootstrap for time series analysis.
}
\examples{

# Example
# generate the original data
block_list = list(c(1:5), c(51:55), c(101:105), c(501:505), c(601:605))
N <- 100; n <- 4; W=10
# X is a list
tp <- unlist(block_list)
tp <- tp - min(tp) + 1
tT <- max(tp)
len <- length(tp)
pop <- seq(N)
group_id <- list()
X <- list()
n <- .data.clean(n, tp)
for (i in tp) {
  group_id[[i]] <- sample(W, N, replace = TRUE) # Individuals randomly form W groups
  g <- sample(W, n[i]) # n groups are randomly selected from W groups
  X[[i]] <- pop[which(group_id[[i]] \%in\% g)] # The members contained in the selected groups
}
# bootstrap sample data
new_sample <- lar_bootstrap(X, block_list, group_id)
# X is a matrix
matrix_data <- list_to_matrix(X, tp, group_id)
new_sample <- lar_bootstrap(matrix_data, block_list)


}
\references{
Politis, D. N. (2003). The impact of bootstrap methods on time series analysis.
Statistical science, 18, 219-230.
}
