% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame_to_matrix.R
\name{data.frame_to_matrix}
\alias{data.frame_to_matrix}
\title{Convert data frame form to matrix form}
\usage{
data.frame_to_matrix(obj)
}
\arguments{
\item{obj}{A 'data.frame' object including ids, observation time, and groups (not required) of individuals}
}
\description{
Convert data frame form to matrix form
}
