% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functions}
\alias{functions}
\alias{.data.clean}
\alias{list_to_matrix}
\alias{data.frame_to_matrix}
\alias{lir_estimation_c}
\alias{lar_estimation_c}
\alias{g_np}
\title{Preparation functions}
\usage{
.data.clean(n, tp)

list_to_matrix(X, tp, group_id = NULL)

data.frame_to_matrix(X)

lir_estimation_c(R_m, R_n, mij, nij, tauij, mtau = 1000)

lar_estimation_c(g_m, g_n, Aij, Ai, tauij, mtau = 1000)

g_np(mij, np, tauij)
}
\arguments{
\item{n}{The number of sampling individuals (in LIR) or groups (in LAR).}

\item{tp}{A set of observed time.}

\item{X}{Simulation data.}

\item{group_id}{Groups of individuals.}

\item{R_m}{Sum of mij.}

\item{R_n}{Sum of nij.}

\item{mij}{The number of individuals identified at both time t_i and t_j.}

\item{nij}{The number of individuals identified at time t_i.}

\item{tauij}{Time lag between time t_i and t_j.}

\item{mtau}{The maximum allowable lag time.}

\item{g_m}{Sum of Aij.}

\item{g_n}{Sum of Ai.}

\item{Aij}{The number of observed associated individuals (in pairs) observed at both time t_i and t_j.}

\item{Ai}{The number of observed associated individuals (in pairs) at time t_i.}

\item{np}{Expectation value of mij (in LIR) or Aij (in LAR).}
}
\description{
Preparation functions
}
