% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_c.R
\name{group_comb}
\alias{group_comb}
\title{Summarize categories}
\usage{
group_comb(real_val, est_val, tauij)
}
\arguments{
\item{real_val}{A series of real values}

\item{est_val}{A series of estimates}

\item{tauij}{Time lag between time t_i and t_j}
}
\description{
Categories were lumped so that all time lag categories contained an expected value of at least six.
}
