% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Nonparametric jackknife resampling for animal movement (social structure) data}
\usage{
jackknife(X, tp, bin_len)
}
\arguments{
\item{X}{A list or matrix containing animal movement or social structure data}

\item{tp}{A set of observation time}

\item{bin_len}{An integer represents len-time-unit intervals}
}
\value{
The jackknife samples of animal movement or social structure data.
}
\description{
jackknife with sampling periods within k-time-unit intervals being omitted
in turn, giving T/k jackknife replicates, with T is the number of sampling periods;
}
\examples{
# Example
# load data
data(simulationA)
tp <- simulationA@tp
# if X is a list
list_simulation_A <- simulationA@list_simulation_A
jackknife_sample <- jackknife(list_simulation_A, tp, bin_len=50)
# if X is a matrix
matrix_simulation_A <- simulationA@matrix_simulation_A
jackknife_sample <- jackknife(matrix_simulation_A, tp, bin_len=50)

}
\references{
Quenouille, M. (1949). Approximate tests of correlation in time series. Journal of the
Royal Statistical Society, Soc. Ser. B, 11, 18-84.

Tukey, J. W. (1958). Bias and confidence in not quite large samples (abstract), Annals of
Mathematical Statistics, 29, 614.

Efron, B. (1979). Bootstrap methods: Another look at the jackknife. Annals of Statistics, 7, 1-26.
}
