% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{lar_bootstrap}
\alias{lar_bootstrap}
\title{Nonparametric block bootstrap sampling for animal social structure data}
\usage{
lar_bootstrap(X, block_list, group_id = NULL, seed = NULL)
}
\arguments{
\item{X}{A list or matrix containing identities of individuals and groups affiliated with them per sampling period}

\item{block_list}{A block list for a series of observation time}

\item{group_id}{Groups to which individuals belong; if X is a list, please input group_id;
if X is a matrix, this parameter can be skipped and and takes the default \code{NULL} value}

\item{seed}{Random seed; the default is NULL}
}
\value{
The bootstrap samples of animal social structure data.
}
\description{
Observation time is divided into non-overlapping blocks; obtain one bootstrap subsample from
each block by sampling the periods with replacement; a complete bootstrap sample is obtained
by merging the bootstrap subsamples according to their order in the original time series.
}
\details{
For more details on this function, please see the block bootstrap sampling for time series analysis.
}
\examples{

# Example
# load data
data(simulationD)
block_list <- simulationD@block_list
group_id <- simulationD@group_id
# if X is a list
list_simulation_D <- simulationD@list_simulation_D
bootstrap_sample <- lar_bootstrap(list_simulation_D, block_list, group_id)
# if X is a matrix
matrix_simulation_D <- simulationD@matrix_simulation_D
bootstrap_sample <- lar_bootstrap(matrix_simulation_D, block_list)



}
\references{
Politis, D. N. (2003). The impact of bootstrap methods on time series analysis.
Statistical science, 18, 219-230.
}
