% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lar_model_selection.R
\name{lar_model_selection}
\alias{lar_model_selection}
\title{Model selection for models of lagged association rate}
\usage{
lar_model_selection(
  X,
  model,
  method,
  tp,
  mtau = 1000,
  ncores = 4,
  nboot = -1,
  k = -1,
  group_id = NULL,
  model_cl_fun = NULL,
  cl.H = NULL,
  model.K = NULL,
  seed = NULL
)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals within
study area, and the states or status of individuals during each sampling period}

\item{model}{Models of lagged identification rate, model = 'lar_1', 'lar_2', 'lar_3', or your model 'model_cl_fun'}

\item{method}{The method = 'Bootstrap', 'BBootstrap', or 'Jackknife'}

\item{tp}{A set of observed time}

\item{mtau}{The maximum allowable lag time}

\item{ncores}{doParallel}

\item{nboot}{The number of bootstrap samples desired}

\item{k}{An integer represents k-time-unit intervals}

\item{group_id}{Groups of individuals. If X is a list, please input group_id. If X is a matrix,
this parameter can be skipped and takes the default \code{NULL} value}

\item{model_cl_fun}{If you formulate your model, please input function to calculate the composite likelihood about your model}

\item{cl.H}{If you formulate your model, please input the sensitivity matrix with respect to parameters in your model}

\item{model.K}{If you formulate your model, please input the number of parameters in your model}

\item{seed}{Random seed}
}
\value{
The values of model selection criteria.
}
\description{
Model selection for models of lagged association rate
}
\details{
See Akaike (1973) for Akaike information criterion (AIC);
See Burnhan and Anderson (2002) for Quasi-Akaike information criterion (QAIC);
See this paper for composite likelihood information criterion (CLIC).
}
