% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lar_nonparametric_estimation.R
\name{lar_nonparametric_estimation}
\alias{lar_nonparametric_estimation}
\title{A nonparametric estimator for lagged association rate}
\usage{
lar_nonparametric_estimation(X, tp, group_id = NULL)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals within
study area, and the states or status of individuals during each sampling period.}

\item{tp}{A set of observation time.}

\item{group_id}{Groups of individuals. If X is a list, please input group_id. If X is a matrix,
this parameter can be skipped and takes the default \code{NULL} value.}
}
\value{
A list with the following elements:
\item{tau}{a set of time lags}
\item{gtau}{A nonparametric estimator of \eqn{\hat{g}(\tau)} was given by Whitehead (2007)}
\item{g_m}{\eqn{\sum_{i,j} \{A_{t_i,t_j}|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|} }
\item{g_n}{\eqn{\sum_{i,j} \{A_i|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|}}
\item{Aij}{number of associations of pair of individuals at both time \eqn{t_i} and \eqn{t_j}}
\item{Ai}{number of associations of pair of individuals at time \eqn{t_i}}
\item{tauij}{\eqn{\tau_{ij}=|t_i-t_j|}}
}
\description{
A nonparametric estimator for lagged association rate
}
\details{
The lagged association rate \eqn{g(\tau)} is the probability that if two individuals
are associated now, they will still be associated after a time lag of  \eqn{\tau}.
A nonparametric estimator of \eqn{g(\tau)} was given by Whitehead (2007):
\deqn{\hat{g}(\tau)=\frac{\sum_{i,j} \{A_{t_i,t_j}|\tau_{ij}=\tau\}}{\sum_{i,j}\{A_i|\tau_{ij}=\tau\}},}
where where \eqn{A_{t_i,t_j}} is the number of associations of pair of individuals at both
time \eqn{t_i} and \eqn{t_j}, and \eqn{A_{t_i}} represents the number of associations
of pair of individuals at time \eqn{t_i}.
}
