% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{lir_bootstrap}
\alias{lir_bootstrap}
\title{Nonparametric bootstrap sampling for animal movement data}
\usage{
lir_bootstrap(X, tp, seed = NULL)
}
\arguments{
\item{X}{A list or matrix containing identities of individuals identified per sampling period}

\item{tp}{A set of observation time}

\item{seed}{Random seed; the default is NULL}
}
\value{
The bootstrap samples of animal movement data.
}
\description{
Generate bootstrap samples from original observations by sampling individuals with replacement.
}
\details{
For more details on this function, please see Efron and Tibshirani (1993).
}
\examples{
# Example
# load data
data(simulationA)
tp <- simulationA@tp
# if X is a list
list_simulation_A <- simulationA@list_simulation_A
bootstrap_sample <- lir_bootstrap(list_simulation_A, tp)
# if X is a matrix
matrix_simulation_A <- simulationA@matrix_simulation_A
bootstrap_sample <- lir_bootstrap(matrix_simulation_A, tp)

}
\references{
Efron, B., & Tibshirani, R. J. (1986). Bootstrap methods for standard errors,
confidence intervals, and other measures of statistical accuracy. Statistical science, 1(1), 54-75.

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. Chapman and Hall Press.
}
