% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_c.R
\name{lir_estimation_c}
\alias{lir_estimation_c}
\title{Estimation of variance inflation factor in QAIC under LIR models}
\usage{
lir_estimation_c(R_m, R_n, mij, nij, tauij, mtau = 1000)
}
\arguments{
\item{R_m}{\eqn{\sum_{i,j} \{m_{t_i,t_j}|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|}}

\item{R_n}{\eqn{\sum_{i,j} \{n_i*n_j|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|}}

\item{mij}{Number of individuals identified at both time t_i and t_j}

\item{nij}{ni*nj where ni is the number of individuals identified at time t_i}

\item{tauij}{Time lag between time t_i and t_j}

\item{mtau}{Maximum allowable lag time}
}
\description{
Estimation of variance inflation factor in QAIC under LIR models
}
