% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lir_model_selection.R
\name{lir_model_selection}
\alias{lir_model_selection}
\title{Model selection for models of lagged identification rate}
\usage{
lir_model_selection(
  X,
  n,
  tp,
  model,
  method,
  ncores = 4,
  mtau = 1000,
  nboot = -1,
  bin_len = -1,
  model_cl_fun = NULL,
  cl.H = NULL,
  model.K = NULL,
  seed = NULL
)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals identified in each sampling period}

\item{n}{A vector or a positive integer, representing the number of individuals identified in each sampling period
It indicates the same number of individuals identified in all sampling periods if a positive integer}

\item{tp}{A set of observed time}

\item{model}{Models of lagged identification rate, model = 'lir_1', 'lir_2', 'lir_3', or formulate model by yourself 'model_cl_fun'}

\item{method}{The method = 'Bootstrap', 'BBootstrap', or 'Jackknife'}

\item{ncores}{doParallel.}

\item{mtau}{The maximum allowable lag time.}

\item{nboot}{The number of bootstrap samples desired}

\item{bin_len}{An integer represents len-time-unit intervals}

\item{model_cl_fun}{If you formulate your model, please input function to calculate the composite likelihood about your model}

\item{cl.H}{If you formulate your model, please input the sensitivity matrix with respect to parameters in your model}

\item{model.K}{If you formulate your model, please input the number of parameters in your model}

\item{seed}{Random seed.}
}
\value{
The values of model selection criteria.
}
\description{
Model selection for models of lagged identification rate
}
\details{
See Akaike (1973) for Akaike information criterion (AIC);
See Burnhan and Anderson (2002) for Quasi-Akaike information criterion (QAIC);
See this paper for composite likelihood information criterion (CLIC).
}
