% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lir_nonparametric_estimation.R
\name{lir_nonparametric_estimation}
\alias{lir_nonparametric_estimation}
\title{A nonparametric estimator for lagged identification rate}
\usage{
lir_nonparametric_estimation(X, n, tp)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals identified during each sampling period}

\item{n}{A vector or a positive integer, representing the number of individuals identified in each sampling period.
It indicates the same number of individuals identified in all sampling periods if a positive integer.}

\item{tp}{A set of observation time}
}
\value{
A list with the following elements:

\item{tau}{a set of time lags}
\item{R_tau}{A nonparametric estimator of \eqn{\hat{R}(\tau)} was given by Whitehead (2007)}
\item{R_m}{\eqn{\sum_{i,j} \{m_{t_i,t_j}|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|} }
\item{R_n}{\eqn{\sum_{i,j} \{n_i*n_j|\tau_{ij}=\tau\}}, with \eqn{\tau_{ij}=|t_i-t_j|}}
\item{mij}{number of individuals identified at both time \eqn{t_i} and \eqn{t_j}}
\item{nij}{\eqn{n_i}*\eqn{n_j}}
\item{tauij}{\eqn{\tau_{ij}=|t_i-t_j|}}
}
\description{
A nonparametric estimator for lagged identification rate
}
\details{
The lagged identification rate \eqn{R(\tau)} is the probability that an individual
in the study area is identified now and is reidentified after a time lag of \eqn{\tau}.
A nonparametric estimator of \eqn{R(\tau)} was given by Whitehead (2007):
\deqn{\hat{R}(\tau)=\frac{\sum_{i,j} \{m_{t_i,t_j}|\tau_{ij}=\tau\}}{\sum_{i,j}\{n_i*n_j|\tau_{ij}=\tau\}},}
where where \eqn{m_{t_i,t_j}} is the number of individuals identified at both
time \eqn{t_i} and \eqn{t_j}, and \eqn{n_{t_i}} represents the number of individuals
identified at time \eqn{t_i}.
}
\examples{
# Example
tp <- c(1:5, 51:55, 101:105, 501:505, 601:605)
N <- 100; n <- 10
X <- list()
for (i in tp){
X[[i]] <- sample(1:N, n)
}

res <- lir_nonparametric_estimation(X, n, tp)
res$R_tau; res$tau


}
