% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_matrix.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Convert a list to matrix form}
\usage{
list_to_matrix(obj, tp, group_id = NULL)
}
\arguments{
\item{obj}{A 'list' object including the ids of individuals}

\item{tp}{A set of observation time}

\item{group_id}{Groups individuals belong to; this parameter can be skipped and takes the default \code{NULL} value}
}
\description{
Convert a list to matrix form
}
