% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_internal.R
\name{models_internal}
\alias{models_internal}
\alias{.mod1}
\alias{.mod2}
\alias{.mod3}
\alias{.mod4}
\alias{.mod5}
\alias{.mod6}
\alias{lir.model.res}
\alias{lar.model.res}
\title{Build the models}
\usage{
.mod1(mij, nij, tauij, mtau = 1000)

.mod2(mij, nij, tauij, mtau = 1000)

.mod3(mij, nij, tauij, mtau = 1000)

.mod4(Aij, Ai, tauij, mtau = 1000)

.mod5(Aij, Ai, tauij, mtau = 1000)

.mod6(Aij, Ai, tauij, mtau = 1000)

lir.model.res(model, mij, nij, tauij, mtau)

lar.model.res(model, Aij, Ai, tauij, mtau)
}
\arguments{
\item{mij}{The number of individuals identified at both time t_i and t_j.}

\item{nij}{The number of individuals identified at time t_i.}

\item{tauij}{Time lag between time t_i and t_j.}

\item{mtau}{The maximum allowable lag time.}

\item{Aij}{The number of observed associated individuals (in pairs) observed at both time t_i and t_j.}

\item{Ai}{The number of observed associated individuals (in pairs) at time t_i.}

\item{model}{Models of LIR and LAR.}
}
\description{
Build the models
}
