% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_A}
\alias{simulation_A}
\title{Type A: Closed}
\format{
contain population size in the study area, number of individuals randomly selected per sampling period,
observation time, list form of simulation A, and matrix form of simulation A
}
\usage{
simulation_A
}
\description{
Type A: Closed
}
\keyword{datasets}
