% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_D}
\alias{simulation_D}
\title{Type D: Random}
\format{
contain population size in the study area, number of groups formed by individuals, number of groups randomly selected per sampling period,
list form of observation time, list form of simulation D, and matrix form of simulation D
}
\usage{
simulation_D
}
\description{
Type D: Random
}
\keyword{datasets}
