% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulationA}
\alias{simulationA}
\title{Type A: Closed}
\format{
an S4 object containing population size in the study area, number of individuals randomly selected per sampling period,
observation time, list form of simulation A, and matrix form of simulation A
}
\usage{
simulationA
}
\description{
Type A: Closed
}
\keyword{datasets}
