% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulationB}
\alias{simulationB}
\title{Type B: Permanent emigration}
\format{
an S4 object containing population size in the study area, number of individuals randomly selected per sampling period,
observation time, rate of emigration, list form of simulation B, and matrix form of simulation B
}
\usage{
simulationB
}
\description{
Type B: Permanent emigration
}
\keyword{datasets}
