% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulationC}
\alias{simulationC}
\title{Type C: Emigration plus reimmigration}
\format{
an S4 object containing population size in the study area, number of individuals randomly selected per sampling period,
observation time, rate of emigration, rate of reimmigration, list form of simulation C, and matrix form of simulation C
}
\usage{
simulationC
}
\description{
Type C: Emigration plus reimmigration
}
\keyword{datasets}
