% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulationE}
\alias{simulationE}
\title{Type E: Casual acquaintances}
\format{
an S4 object containing population size in the study area, number of groups formed by individuals, number of groups randomly selected per sampling period,
probability of individual changes group per time unit, list form of observation time, list form of simulation E, and matrix form of simulation E
}
\usage{
simulationE
}
\description{
Type E: Casual acquaintances
}
\keyword{datasets}
