% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulationF}
\alias{simulationF}
\title{Type F: Permanent companions plus casual acquaintances}
\format{
an S4 object containing population size in the study area, number of groups formed by individuals, Number of permanent social units allocated by individuals,
number of groups randomly selected per sampling period, probability of individual changes group per time unit, list form of observation time, list form of simulation F, and matrix form of simulation F
}
\usage{
simulationF
}
\description{
Type F: Permanent companions plus casual acquaintances
}
\keyword{datasets}
