% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_ci.R
\name{width_ci}
\alias{width_ci}
\title{Compute the width of confidence intervals}
\usage{
width_ci(x, alpha = 0.05)
}
\arguments{
\item{x}{A set of estimated parameters}

\item{alpha}{Confidence level is 1-alpha}
}
\value{
The width of confidence intervals.
}
\description{
The width is the range between the 2.5 and 97.5\% percentiles of the estimated parameters for many (such as 100) random runs
}
\examples{
x <- rnorm(100)
width_ci(x)

}
