% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lar_model_selection.R
\name{lar_model_selection}
\alias{lar_model_selection}
\title{Model selection for models of lagged association rate}
\usage{
lar_model_selection(
  X,
  model,
  block_list,
  nboot,
  group_id = NULL,
  mtau = 1000,
  ncores = 4,
  seed = NULL
)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals within
study area, and the states or status of individuals during each sampling period.}

\item{model}{Models of lagged identification rate, model = 'lar_1', 'lar_2', or 'lar_3'.}

\item{block_list}{A block list for a series of observation time. For example,
block_list = list(c(1:5), c(51:55), c(101:105), c(501:505), c(601:605)).}

\item{nboot}{The number of bootstrap samples desired}

\item{group_id}{Groups of individuals. If X is a list, please input group_id. If X is a matrix,
this parameter can be skipped and takes the default \code{NULL} value.}

\item{mtau}{The maximum allowable lag time}

\item{ncores}{doParallel}

\item{seed}{Random seed}
}
\value{
The values of model selection criteria.
}
\description{
Model selection for models of lagged association rate
}
\details{
See Akaike (1973) for Akaike information criterion (AIC);
See Burnhan and Anderson (2002) for Quasi-Akaike information criterion (QAIC);
See this paper for composite likelihood information criterion (CLIC).
}
