% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lir_model_selection.R
\name{lir_model_selection}
\alias{lir_model_selection}
\title{Model selection for models of lagged identification rate}
\usage{
lir_model_selection(
  X,
  n,
  tp,
  model,
  nboot,
  mtau = 1000,
  ncores = 4,
  seed = NULL
)
}
\arguments{
\item{X}{A list or matrix containing the identities of individuals identified in each sampling period}

\item{n}{A vector or a positive integer, representing the number of individuals identified in each sampling period.
It indicates the same number of individuals identified in all sampling periods if a positive integer.}

\item{tp}{A set of observed time.}

\item{model}{Models of lagged identification rate, model = 'lir_1', 'lir_2', or 'lir_3'.}

\item{nboot}{The number of bootstrap samples desired}

\item{mtau}{The maximum allowable lag time.}

\item{ncores}{doParallel.}

\item{seed}{Random seed.}
}
\value{
The values of model selection criteria.
}
\description{
Model selection for models of lagged identification rate
}
\details{
See Akaike (1973) for Akaike information criterion (AIC);
See Burnhan and Anderson (2002) for Quasi-Akaike information criterion (QAIC);
See this paper for composite likelihood information criterion (CLIC).
}
